# Faster strpos() string function for ASCII-only case