@goldmedal 2024-09-11
> I think it's not executed and triggered some error, see benchmark `StringArray` cases only take several ns, and `StringViewArray` cases are at least several us I guess it's because in micro bench, arguments are of different physical string type and break the function implementation (`strpos(string_col, string_view_col)`)

You're right. There are some errors here. I ran the benchmark in the `strpos` test case and got the following error:
```
Error: Execution("Unsupported data type combination (Utf8, Utf8View) for function strpos")
```
It seems that `strpos` doesn't allow the second argument to be a `StringView` if the first one is a `StringArray`. I'm not sure if this case makes sense. Maybe we can file another issue for it if it's reasonable.

Then, I realized it was my mistake 😢. When I generated the substring for the `StringArray`, I mistakenly set its type to `StringView`. After fixing it, the benchmark for `StringArray` works fine.