@goldmedal 2024-09-11
@2010YOUY01 Thank you for the review. After fixing the benchmark, we can find the StringArray case also be improved.